+++ lib/flatten.cpp
@@ -906,6 +906,7 @@ EnvI::EnvI(Model* model0, std::ostream& outstream0, std::ostream& errstream0)
       inMaybePartial(0),
       inTraceExp(false),
       inReverseMapVar(false),
+      warnImplicitEnum2Int(true),
       counters({0, 0, 0, 0}),
       _flat(new Model),
       _failed(false),