+++ lib/model.cpp
@@ -1065,7 +1065,8 @@ int first_overloaded(EnvI& env, const std::vector<Model::FnEntry>& v_f, int i_f)
 
 bool Model::sameOverloading(EnvI& env, const std::vector<Expression*>& args, FunctionI* f,
                             FunctionI* g) const {
-  if (f->id() == env.constants.varRedef->id() || g->id() == env.constants.varRedef->id()) {
+  if (f->isMonomorphised() || g->isMonomorphised() || f->id() == env.constants.varRedef->id() ||
+      g->id() == env.constants.varRedef->id()) {
     return false;
   }
   const Model* m = this;